[MRG] Fix deprecated np.bool, np.int and np.float and array of inhomogeneous shape (#469)

+++ tslearn/tests/test_serialize_models.py
@@ -38,8 +38,8 @@ def clear_tmp():
 
 
 def test_hdftools():
-    dtypes = [numpy.int, numpy.int8, numpy.int16, numpy.int32, numpy.int64,
-              numpy.float, numpy.float32, numpy.float64]
+    dtypes = [int, numpy.int8, numpy.int16, numpy.int32, numpy.int64,
+              float, numpy.float32, numpy.float64]
 
     d = {}
